readme: Update Galileo installation instructions
Remove the information that tells user to dd the image to SD card.
This does not work properly, like console is missing etc. in
Galileo when booting. We need to manually copy relevant files into
SD card.

A link to proper kernel configuration files need to be setup
manually by the user.